# public/openapi.yaml
openapi: "3.1.0"
servers:
  - url: https://ccapdev-mco-pq20.onrender.com/api
info:
  version: "0.1.0"
  title: "ArchersLab API"
  description: CCAPDEV MCO
  license:
    name: MIT
    identifier: MIT
security:
  - Jwt: []
  - {}
paths:
  /login:
    post:
      summary: Login
      operationId: login
      tags:
        - user
      responses:
        "204":
          description: Create session
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/Unexpected"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserLogin"
    delete:
      summary: Logout
      operationId: logout
      tags:
        - user
      security:
        - Jwt: []
      responses:
        "204":
          description: Logs the user out
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/Unexpected"
...
  /lab:
    get:
      summary: Read all labs
      operationId: readLabs
      tags:
        - lab
      parameters:
        - name: q
          description: The search query to filter labs
          in: query
          schema:
            type: string
        - name: page
          description: Pagination query
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
      responses:
        "200":
          description: The labs containing the provided `q` in name
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Lab"
                  hasNextPage:
                    $ref: "#/components/schemas/HasNextPage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/Unexpected"
    post:
      summary: Create a lab
      operationId: createLab
      tags:
        - lab
      security:
        - Jwt: []
      responses:
        "201":
          description: The lab created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabId"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Exists"
        "500":
          $ref: "#/components/responses/Unexpected"
      requestBody:
        description: The lab to create
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LabDetails"
  /lab/{id}:
    parameters:
      - $ref: "#/components/parameters/labId"
    get:
      summary: Read a lab
      operationId: readLab
      tags:
        - lab
      responses:
        "200":
          description: The lab corresponding to the provided `id`
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lab"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/Unexpected"
    put:
      summary: Update a lab
      operationId: updateLab
      tags:
        - lab
      security:
        - Jwt: []
      responses:
        "204":
          description: Update the lab corresponding to the provided `id`
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/Unexpected"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LabDetails"
    delete:
      summary: Delete a lab
      operationId: deleteLab
      tags:
        - lab
      security:
        - Jwt: []
      responses:
        "204":
          description: Delete the lab corresponding to the provided `id`
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/Unexpected"
...
  /reservation/{id}:
    parameters:
      - name: id
        description: The unique identifier of the reservation
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/ReservationId"
    put:
      summary: Update a reservation
      operationId: updateReservation
      tags:
        - reservation
      security:
        - Jwt: []
      responses:
        "204":
          description: Update the reservation corresponding to the provided `id`
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/Unexpected"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReservationUpdate"
    delete:
      summary: Delete a reservation
      operationId: deleteReservation
      tags:
        - reservation
      security:
        - Jwt: []
      responses:
        "204":
          description: Delete the reservation corresponding to the provided `id`
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/Unexpected"
...
components:
  schemas:
    Id:
      description: ObjectID
      type: string
      pattern: "^[0-9a-fA-F]{24}$"
      example: "507f1f77bcf86cd799439011"
    Editable:
      type: boolean
    HasNextPage:
      type: boolean
    Interval:
      description: Time interval
      type: object
      required:
        - start
        - end
      properties:
        start:
          type: string
          format: date-time
          example: "1970-01-01T00:00:00.000Z"
        end:
          type: string
          format: date-time
          example: "1970-01-01T00:00:00.000Z"
    SlotId:
      type: string
      minLength: 1
...
    LabId:
      description: The unique identifier of a lab
      $ref: "#/components/schemas/Id"
      example: 1
    ReservationId:
      description: The unique identifier of a reservation
      $ref: "#/components/schemas/Id"
      example: 1
...
    UserPassword:
      type: string
      format: password
      example: "123456"
      writeOnly: true
    UserLogin:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          $ref: "#/components/schemas/UserPassword"
...
    LabDetails:
      type: object
      required:
        - name
        - weeklySchedule
        - slots
      properties:
        name:
          type: string
          example: "GK403"
        weeklySchedule:
          description: Weekly schedule of the lab
          type: object
          properties:
            sunday:
              $ref: "#/components/schemas/Interval"
            monday:
              $ref: "#/components/schemas/Interval"
            tuesday:
              $ref: "#/components/schemas/Interval"
            wednesday:
              $ref: "#/components/schemas/Interval"
            thursday:
              $ref: "#/components/schemas/Interval"
            friday:
              $ref: "#/components/schemas/Interval"
            saturday:
              $ref: "#/components/schemas/Interval"
        slots:
          description: Available slots in the lab
          type: array
          items:
            type: object
            required:
              - id
              - x
              - y
            properties:
              id:
                $ref: "#/components/schemas/SlotId"
              x:
                type: integer
                minimum: 0
              y:
                type: integer
                minimum: 0
            example:
              id: "1"
              x: 0
              y: 0
    Lab:
      allOf:
        - $ref: "#/components/schemas/LabDetails"
        - type: object
          required:
            - id
          properties:
            id:
              $ref: "#/components/schemas/LabId"
            editable:
              $ref: "#/components/schemas/Editable"
    ReservationUpdate:
      type: object
      required:
        - schedule
        - slotIds
      properties:
        anonymous:
          type: boolean
        schedule:
          $ref: "#/components/schemas/Interval"
        slotIds:
          type: array
          items:
            $ref: "#/components/schemas/SlotId"
...
    Error:
      type: object
      required:
        - message
      properties:
        message:
          description: A human readable error message
          type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: No items found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Unauthorized access
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Exists:
      description: Data already Exists
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unexpected:
      description: Unexpected error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  parameters:
...
    labId:
      name: id
      description: The unique identifier of the lab
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/LabId"
  securitySchemes:
    Jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT