# src/api/models/reservationUpdate.ts
/**
 * Generated by orval v8.3.0 🍺
 * Do not edit manually.
 * ArchersLab API
 * CCAPDEV MCO
 * OpenAPI spec version: 0.1.0
 */
import type { Interval } from './interval';
import type { SlotId } from './slotId';

export interface ReservationUpdate {
  anonymous?: boolean;
  schedule: Interval;
  slotIds: SlotId[];
}